src/app/cetificate-authorities/details/CertificateAuthorityDetailsClient
